chore: resolve obj store dep update and update hakari
Updating hakari required some manual changes to Cargo.toml files to
bring in features for some dependencies. This is a bit odd, because now,
when a sync from IOx is done, this may get overwritten. It is likely that
IOx has the features for these crates specified in some other, non-core
crate, so the compiler errors are not seen there.