chore(cargo): remove workspace hack from hakari
Workspace hack is appropriate in projects building binaries. This is a
project full of shared crates by other projects. This project may build
a slower without workspace-hack but dependent projects should be able to
better manage their deps, which might be a subset of core, without
pulling in every single crate through the workspace-hack dep.